Conversations module was finished

+++ src/conversations/conversations.service.ts
@@ -1,6 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
+import { LoggerWinston } from '../logger/logger-winston.service';
+import { User, UserDocument } from '../users/user.schema';
 import { UsersService } from '../users/users.service';
 import { Conversation, ConversationDocument } from './conversation.schema';
 import { CreateConversationDto } from './dto/create-conversation.dto';
@@ -11,20 +13,22 @@ export class ConversationsService {
   constructor(
     @InjectModel(Conversation.name)
     private conversationModel: Model<ConversationDocument>,
+    private readonly logger: LoggerWinston,
+    @InjectModel(User.name)
+    private userModel: Model<UserDocument>,
     private readonly usersService: UsersService,
   ) {}
 
   async create(createConversationDto: CreateConversationDto, userId: string): Promise<Conversation> {
     const { name, interlocutorId } = createConversationDto;
-    const createdConversation = new this.conversationModel({
-      name,
-      users: {},
-    });
 
     const user = await this.usersService.findById(userId);
-    createdConversation.users.set(interlocutorId, user);
     const interlocutor = await this.usersService.findById(interlocutorId);
-    createdConversation.users.set(userId, interlocutor);
+
+    const createdConversation = new this.conversationModel({
+      name,
+      users: new Map([[userId, interlocutor], [interlocutorId, user]]),
+    });
 
     user.myConversations.push(createdConversation._id);
     await this.usersService.update(userId, user);
@@ -35,19 +39,41 @@ export class ConversationsService {
     return createdConversation.save();
   }
 
-  findAll(): Promise<Conversation[]> {
-    return this.conversationModel.find({}).exec();
+  async findAll(userId: string): Promise<(Conversation | string)[]> {
+    const user = await this.userModel.findById(userId).populate('myConversations', '_id name unreadMessageFromUser users').exec();
+    return user.myConversations;
   }
 
-  findOne(id: string): Promise<Conversation> {
-    return this.conversationModel.findById(id).populate('messages').exec();
+  async findOne(id: string): Promise<Conversation> {
+    const conversation = await this.conversationModel.findById(id).populate('messages').exec();
+
+    if (!conversation) {
+      this.logger.warn(`User error: Conversation with id = ${id} not found.`, 'ConversationsService');
+      throw new NotFoundException();
+    }
+
+    return conversation;
   }
 
   update(id: string, updateConversationDto: UpdateConversationDto): Promise<Conversation> {
     return this.conversationModel.findByIdAndUpdate(id, updateConversationDto, { new: true }).exec();
   }
 
-  remove(id: string): Promise<Conversation> {
-    return this.conversationModel.findByIdAndDelete(id).exec();
+  async remove(id: string, userId: string): Promise<Conversation> {
+    const conversation = await this.conversationModel.findById(id).exec();
+
+    /* // Socket ?
+
+    if (conversation.users.size < 2) {
+      return this.conversationModel.findByIdAndDelete(id).exec();
+    } */
+
+    conversation.users.delete(userId);
+
+    const user = await this.userModel.findById(userId).exec();
+    user.myConversations = user.myConversations.filter(conversationId => conversationId.toString() !== id);
+    await this.usersService.update(userId, user);
+
+    return this.update(id, conversation);
   }
 }
